baked lightmaps for the mall, added occlusion culling, and did a ton of changes with materials to make them look better with baked lights

+++ Assets/mats/ChainLink/OptimizedShaders/Chain/Poiyomi.shader
@@ -2468,7 +2468,7 @@ Shader "Hidden/Locked/.poiyomi/Poiyomi 8.0/Poiyomi Toon/cfed4a7b442967c43932e180
 				reflectionMask *= PRBMaps.b;
 				specularMask *= PRBMaps.a;
 				#endif
-				reflectionMask *= float(1);
+				reflectionMask *= float(0);
 				specularMask *= float(0.678);
 				if (float(0))
 				{
@@ -5137,7 +5137,7 @@ Shader "Hidden/Locked/.poiyomi/Poiyomi 8.0/Poiyomi Toon/cfed4a7b442967c43932e180
 				reflectionMask *= PRBMaps.b;
 				specularMask *= PRBMaps.a;
 				#endif
-				reflectionMask *= float(1);
+				reflectionMask *= float(0);
 				specularMask *= float(0.678);
 				if (float(0))
 				{
